FIX: Changed double to BigDecimal for financial accuracy

+++ app/src/main/java/com/neueda/app/dto/AccountResponse.java
@@ -1,5 +1,7 @@
 package com.neueda.app.dto;
 
+import java.math.BigDecimal;
+
 import com.neueda.app.enums.AccountStatus;
 import lombok.Data;
 
@@ -7,7 +9,7 @@ import lombok.Data;
 public class AccountResponse {
     
     private Long accountID; 
-    private double cash_balance;
+    private BigDecimal cash_balance;
     private AccountStatus status;
     
 }
